Tried to fix issues with non-crop blocks not adhering to seasonal crop listings

+++ src/main/java/sereneseasons/handler/season/SeasonalCropGrowthHandler.java
@@ -2,12 +2,9 @@ package sereneseasons.handler.season;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockCocoa;
-import net.minecraft.block.BlockDoublePlant;
 import net.minecraft.block.BlockGrass;
-import net.minecraft.block.BlockMushroom;
+import net.minecraft.block.BlockReed;
 import net.minecraft.block.BlockSapling;
-import net.minecraft.block.BlockTallGrass;
-import net.minecraft.block.IGrowable;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 import net.minecraftforge.event.entity.player.BonemealEvent;
@@ -38,7 +35,7 @@ public class SeasonalCropGrowthHandler
 		
 		if (isFertilityApplicable(plant) && !isFertile && !isGreenhouseGlassAboveBlock(event.getWorld(), event.getPos()))
 		{
-			if (FertilityConfig.general_category.crops_break)
+			if (FertilityConfig.general_category.crops_break && !(plant instanceof BlockGrass) && !(plant instanceof BlockReed))
 			{
 				event.getWorld().destroyBlock(event.getPos(), true);
 			}
@@ -50,14 +47,14 @@ public class SeasonalCropGrowthHandler
 	}
 
 	@SubscribeEvent
-	public void onBonemeal(BonemealEvent event)
+	public void onApplyBonemeal(BonemealEvent event)
 	{
 		Block plant = event.getBlock().getBlock();
 		boolean isFertile = ModFertility.isCropFertile(plant.getRegistryName().toString(), event.getWorld(), event.getPos());
 		
 		if (isFertilityApplicable(plant) && !isFertile && !isGreenhouseGlassAboveBlock(event.getWorld(), event.getPos()))
 		{
-			if (FertilityConfig.general_category.crops_break)
+			if (FertilityConfig.general_category.crops_break && !(plant instanceof BlockGrass) && !(plant instanceof BlockReed))
 			{
 				event.getWorld().destroyBlock(event.getPos(), true);
 			}
@@ -73,20 +70,13 @@ public class SeasonalCropGrowthHandler
 			return false;
 		}
 		
-		if (!(block instanceof IGrowable))
+		if (block instanceof BlockSapling)
 		{
 			return false;
 		}
 		else
 		{
-			if (block instanceof BlockCocoa)
-			{
-				return false;
-			}
-			else
-			{
-				return true;
-			}
+			return true;
 		}
 	}
